requirements.txt
asgiref==3.8.1
Django==5.0.6
django-ckeditor==6.7.1
django-extensions==3.2.3
django-js-asset==2.2.0
django-mdeditor==0.1.20
Markdown==3.6
sqlparse==0.5.0
tzdata==2024.1
Unidecode==1.3.8
